Docs: "upgrade" to "upgrade-miniapp"
Command "upgrade" does not exist, I believe you meant "upgrade-miniapp"?

That command is supposed to upgrade used ERN platform, but as a side-effect it also upgrades default supported platform modules, too, which is the issue being discussed here. However that is not enough, if you are using some 3rd party or own custom modules.